[autodoc/autotest] ApplyGateToLthQubit
