+++ qualtran/bloqs/apply_gate_to_lth_target.py
@@ -20,7 +20,7 @@ import cirq
 import numpy as np
 from cirq._compat import cached_property
 
-from qualtran import Register, SelectionRegister
+from qualtran import bloq_example, BloqDocSpec, Register, SelectionRegister, Signature
 from qualtran._infra.gate_with_registers import total_bits
 from qualtran.bloqs.unary_iteration_bloq import UnaryIterationGate
 
@@ -39,14 +39,13 @@ class ApplyGateToLthQubit(UnaryIterationGate):
     `selection`-th qubit of `target` all controlled by the `control` register.
 
     Args:
-        selection_regs: Indexing `select` signature of type Tuple[`SelectionRegisters`, ...].
+        selection_regs: Indexing `select` signature of type Tuple[`SelectionRegister`, ...].
             It also contains information about the iteration length of each selection register.
         nth_gate: A function mapping the composite selection index to a single-qubit gate.
         control_regs: Control signature for constructing a controlled version of the gate.
 
     References:
-            [Encoding Electronic Spectra in Quantum Circuits with Linear T Complexity]
-        (https://arxiv.org/abs/1805.03662).
+        [Encoding Electronic Spectra in Quantum Circuits with Linear T Complexity](https://arxiv.org/abs/1805.03662).
         Babbush et. al. (2018). Section III.A. and Figure 7.
     """
     selection_regs: Tuple[SelectionRegister, ...] = attr.field(
@@ -102,3 +101,28 @@ class ApplyGateToLthQubit(UnaryIterationGate):
         selection_idx = tuple(selection_indices[reg.name] for reg in self.selection_regs)
         target_idx = int(np.ravel_multi_index(selection_idx, selection_shape))
         return self.nth_gate(*selection_idx).on(target[target_idx]).controlled_by(control)
+
+
+@bloq_example
+def _apply_z_to_odd() -> ApplyGateToLthQubit:
+    from qualtran import SelectionRegister
+
+    def _z_to_odd(n: int):
+        if n % 2 == 1:
+            return cirq.Z
+        return cirq.I
+
+    apply_z_to_odd = ApplyGateToLthQubit(
+        SelectionRegister('selection', 3, 4),
+        nth_gate=_z_to_odd,
+        control_regs=Signature.build(control=2),
+    )
+
+    return apply_z_to_odd
+
+
+_APPLYLTH_DOC = BloqDocSpec(
+    bloq_cls=ApplyGateToLthQubit,
+    import_line='from qualtran.bloqs.apply_gate_to_lth_target import ApplyGateToLthQubit',
+    examples=(_apply_z_to_odd,),
+)
